Switch to creekservice
part of https://github.com/creek-service/creek-test/issues/24
