+++ annotation/src/test/java/org/creekservice/api/base/annotation/schema/GeneratesSchemaTest.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2022 Creek Contributors (https://github.com/creek-service)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.creekservice.api.base.annotation.schema;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.notNullValue;
+
+import org.junit.jupiter.api.Test;
+
+class GeneratesSchemaTest {
+
+    @Test
+    void shouldRetainAnnotationAtRuntime() {
+        assertThat(
+                TestType.class.getAnnotation(
+                        org.creekservice.api.base.annotation.schema.GeneratesSchema.class),
+                is(notNullValue()));
+    }
+
+    @GeneratesSchema
+    private static final class TestType {}
+}
