+++ schema/src/main/java/module-info.java
@@ -2,6 +2,6 @@ module creek.base.schema {
     requires creek.base.annotation;
     requires io.github.classgraph;
 
-    exports org.creek.api.base.schema;
-    exports org.creek.api.base.schema.naming;
+    exports org.creekservice.api.base.schema;
+    exports org.creekservice.api.base.schema.naming;
 }
