+++ type/src/main/java/module-info.java
@@ -1,8 +1,8 @@
 module creek.base.type {
     requires transitive creek.base.annotation;
 
-    exports org.creek.api.base.type;
-    exports org.creek.api.base.type.config;
-    exports org.creek.api.base.type.temporal;
-    exports org.creek.api.base.type.json;
+    exports org.creekservice.api.base.type;
+    exports org.creekservice.api.base.type.config;
+    exports org.creekservice.api.base.type.temporal;
+    exports org.creekservice.api.base.type.json;
 }
