--- type/src/main/java/org/creek/api/base/type/Lists.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2022 Creek Contributors (https://github.com/creek-service)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.creek.api.base.type;
-
-
-import java.util.ArrayList;
-import java.util.List;
-
-public final class Lists {
-
-    private Lists() {}
-
-    public static <T> List<T> combineList(final List<? extends T> a, final List<? extends T> b) {
-        final List<T> result = new ArrayList<>(a);
-        result.addAll(b);
-        return result;
-    }
-}
