--- type/src/main/java/org/creek/api/base/type/temporal/AccurateClock.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2022 Creek Contributors (https://github.com/creek-service)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.creek.api.base.type.temporal;
-
-
-import java.time.Instant;
-
-/** Default clock impl. */
-public final class AccurateClock implements Clock {
-
-    public static Clock create() {
-        return new AccurateClock();
-    }
-
-    @Override
-    public Instant get() {
-        return Instant.now();
-    }
-
-    private AccurateClock() {}
-}
